Remove `include/grpcpp/impl/codegen/stub_options.h`
