+++ include/grpcpp/impl/codegen/stub_options.h
@@ -19,26 +19,9 @@
 #ifndef GRPCPP_IMPL_CODEGEN_STUB_OPTIONS_H
 #define GRPCPP_IMPL_CODEGEN_STUB_OPTIONS_H
 
-// IWYU pragma: private, include <grpcpp/support/stub_options.h>
+// IWYU pragma: private
 
-namespace grpc {
-
-/// Useful interface for generated stubs
-class StubOptions {
- public:
-  StubOptions() = default;
-  explicit StubOptions(const char* suffix_for_stats)
-      : suffix_for_stats_(suffix_for_stats) {}
-
-  void set_suffix_for_stats(const char* suffix_for_stats) {
-    suffix_for_stats_ = suffix_for_stats;
-  }
-  const char* suffix_for_stats() const { return suffix_for_stats_; }
-
- private:
-  const char* suffix_for_stats_ = nullptr;
-};
-
-}  // namespace grpc
+/// TODO(chengyuc): Remove this file after solving compatibility.
+#include <grpcpp/support/stub_options.h>
 
 #endif  // GRPCPP_IMPL_CODEGEN_STUB_OPTIONS_H
