+++ src/compiler/cpp_generator.cc
@@ -152,7 +152,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/impl/codegen/server_context.h",
         "grpcpp/impl/codegen/service_type.h",
         "grpcpp/impl/codegen/status.h",
-        "grpcpp/impl/codegen/stub_options.h",
+        "grpcpp/support/stub_options.h",
         "grpcpp/impl/codegen/sync_stream.h",
     };
     std::vector<std::string> headers(headers_strs, array_end(headers_strs));
