+++ test/cpp/codegen/compiler_test_golden
@@ -42,7 +42,7 @@
 #include <grpcpp/impl/codegen/server_context.h>
 #include <grpcpp/impl/codegen/service_type.h>
 #include <grpcpp/impl/codegen/status.h>
-#include <grpcpp/impl/codegen/stub_options.h>
+#include <grpcpp/support/stub_options.h>
 #include <grpcpp/impl/codegen/sync_stream.h>
 
 namespace grpc {
